회원리스트, 카드관리 페이지 권한 검증 강화

+++ src/main/java/zgoo/cpos/service/MemberService.java
@@ -46,6 +46,7 @@ import zgoo.cpos.repository.member.MemberCreditCardRepository;
 import zgoo.cpos.repository.member.MemberRepository;
 import zgoo.cpos.util.AESUtil;
 import zgoo.cpos.util.EncryptionUtils;
+import zgoo.cpos.util.MenuConstants;
 
 @Service
 @RequiredArgsConstructor
@@ -61,6 +62,7 @@ public class MemberService {
     public final ConditionCodeRepository conditionCodeRepository;
     public final MemberAuthRepository memberAuthRepository;
     public final ConditionVersionHistRepository conditionVersionHistRepository;
+    public final ComService comService;
 
     // 회원 조회
     public Page<MemberListDto> findMemberInfoWithPagination(Long companyId, String idTag, String name, int page,
@@ -150,8 +152,15 @@ public class MemberService {
 
     // 회원 저장
     @Transactional
-    public void saveMember(MemberRegDto dto) {
+    public void saveMember(MemberRegDto dto, String loginUserId) {
         try {
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot save member info without login user ID."); 
+            }
+
+            boolean isMod = comService.checkModYn(loginUserId, MenuConstants.MEMBER_LIST);
+            if (!isMod) return;
+
             Company company = this.companyRepository.findById(dto.getCompanyId())
                     .orElseThrow(
                             () -> new IllegalArgumentException("company not found with id: " + dto.getCompanyId()));
@@ -198,6 +207,9 @@ public class MemberService {
             // 회원카드관리 - 기본값 설정
             MemberAuth auth = MemberMapper.toEntityAuth(saved);
             this.memberAuthRepository.save(auth);
+        } catch (IllegalArgumentException e) {
+            log.error("[saveMember] Illegal argument error: {}", e.getMessage());
+            throw e;
         } catch (Exception e) {
             log.error("[saveMember] error: {}", e.getMessage());
         }
@@ -261,10 +273,16 @@ public class MemberService {
 
     // 회원 리스트 수정
     @Transactional
-    public Member updateMemberInfo(Long memberId, MemberRegDto dto) {
+    public Member updateMemberInfo(Long memberId, MemberRegDto dto, String loginUserId) {
         Member member = this.memberRepository.findById(memberId)
                 .orElseThrow(() -> new IllegalArgumentException("member not found with id: " + memberId));
         try {
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot update member info without login user ID."); 
+            }
+            boolean isMod = comService.checkModYn(loginUserId, MenuConstants.MEMBER_LIST);
+            if (!isMod) return member;
+
             member.updateMemberInfo(dto);
 
             // 결제카드 정보 수정
@@ -289,6 +307,9 @@ public class MemberService {
 
             updateCarInfo(dto, member);
             updateMemberConditionInfo(dto, member);
+        } catch (IllegalArgumentException e) {
+            log.error("[updateMemberInfo] Illegal argument error: {}", e.getMessage());
+            throw e;
         } catch (Exception e) {
             log.error("[updateMemberInfo] error: {}", e.getMessage());
         }
@@ -435,16 +456,26 @@ public class MemberService {
 
     // 회원 리스트 삭제
     @Transactional
-    public void deleteMember(Long memberId) {
+    public void deleteMember(Long memberId, String loginUserId) {
         Member member = this.memberRepository.findById(memberId)
                 .orElseThrow(() -> new IllegalArgumentException("member not found with id: " + memberId));
 
         try {
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot delete member info without login user ID."); 
+            }
+
+            boolean isMod = comService.checkModYn(loginUserId, MenuConstants.MEMBER_LIST);
+            if (!isMod) return;
+
             this.memberCreditCardRepository.deleteAllByMemberId(memberId);
             this.memberCarRepository.deleteAllByMemberId(memberId);
             this.memberConditionRepository.deleteAllByMemberId(memberId);
             this.memberRepository.delete(member);
             log.info("==== memberId: {} is deleted..", memberId);
+        } catch (IllegalArgumentException e) {
+            log.error("[deleteMember] Illegal argument error: {}", e.getMessage());
+            throw e;
         } catch (Exception e) {
             log.error("[deleteMember] error: {}", e.getMessage());
         }
@@ -558,11 +589,21 @@ public class MemberService {
 
     // 회원카드정보 수정
     @Transactional
-    public MemberAuth updateMemberAuthInfo(MemberAuthDto dto) {
+    public MemberAuth updateMemberAuthInfo(MemberAuthDto dto, String loginUserId) {
         try {
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot update member tag info without login user ID."); 
+            }
+
+            boolean isMod = comService.checkModYn(loginUserId, MenuConstants.MEMBER_TAG);
+            if (!isMod) return null;
+
             MemberAuth memberAuth = this.memberAuthRepository.findByIdTag(dto.getIdTag());
             memberAuth.updateMemberAuthInfo(dto);
             return memberAuth;
+        } catch (IllegalArgumentException e) {
+            log.error("[updateMemberAuthInfo] Illegal argument error: {}", e.getMessage());
+            throw e;
         } catch (Exception e) {
             log.error("[updateMemberAuthInfo] error: {}", e.getMessage());
             return null;
@@ -571,11 +612,21 @@ public class MemberService {
 
     // 회원카드정보 삭제
     @Transactional
-    public void deleteMemberAuth(String idTag) {
+    public void deleteMemberAuth(String idTag, String loginUserId) {
         try {
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot delete member tag info without login user ID."); 
+            }
+
+            boolean isMod = comService.checkModYn(loginUserId, MenuConstants.MEMBER_TAG);
+            if (!isMod) return;
+
             MemberAuth memberAuth = this.memberAuthRepository.findByIdTag(idTag);
             this.memberAuthRepository.delete(memberAuth);
             log.info("==== idTag: {} is deleted..", idTag);
+        } catch (IllegalArgumentException e) {
+            log.error("[deleteMemberAuth] Illegal argument error: {}", e.getMessage());
+            throw e;
         } catch (Exception e) {
             log.error("[deleteMemberAuth] error: {}", e.getMessage());
         }
@@ -600,7 +651,7 @@ public class MemberService {
             }
 
         } catch (Exception e) {
-            e.printStackTrace();
+            log.error("[readDocxFile] error: {}", e.getMessage());
             throw new RuntimeException("파일을 읽는 도중 오류가 발생했습니다.");
         }
